Added login_required to secure views

+++ products/views.py
@@ -2,6 +2,8 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
+
 
 from .models import Product, Category
 from .forms import ProductForm
@@ -73,9 +75,13 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have access to add products.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -97,8 +103,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have access to edit products.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -121,7 +132,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
     
 
+@login_required
 def delete_product(request, product_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have access to delete products.')
+        return redirect(reverse('home'))
+        
     """ Delete a product from the store """
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
